обработка owner
