+++ package.json
@@ -16,6 +16,7 @@
     "express": "4.17.1",
     "helmet": "^4.6.0",
     "joi": "^17.5.0",
+    "joi-objectid": "^4.0.2",
     "joi-phone-number": "^5.0.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.1.1",
